Pal 394 updated classes after moving Policy to Palisade-common
