+++ policy-service/src/main/java/uk/gov/gchq/palisade/service/policy/service/NullPolicyService.java
@@ -19,7 +19,7 @@ package uk.gov.gchq.palisade.service.policy.service;
 import uk.gov.gchq.palisade.Context;
 import uk.gov.gchq.palisade.User;
 import uk.gov.gchq.palisade.resource.Resource;
-import uk.gov.gchq.palisade.service.policy.request.Policy;
+import uk.gov.gchq.palisade.service.request.Policy;
 
 import java.util.Optional;
