+++ policy-service/src/main/java/uk/gov/gchq/palisade/service/policy/service/PolicyServiceCachingProxy.java
@@ -26,7 +26,7 @@ import org.springframework.cache.annotation.Cacheable;
 import uk.gov.gchq.palisade.Context;
 import uk.gov.gchq.palisade.User;
 import uk.gov.gchq.palisade.resource.Resource;
-import uk.gov.gchq.palisade.service.policy.request.Policy;
+import uk.gov.gchq.palisade.service.request.Policy;
 
 import java.util.Optional;
