+++ policy-service/src/test/java/uk/gov/gchq/palisade/service/policy/service/PolicyServiceHierarchyProxyTest.java
@@ -35,7 +35,7 @@ import uk.gov.gchq.palisade.resource.impl.SystemResource;
 import uk.gov.gchq.palisade.rule.PredicateRule;
 import uk.gov.gchq.palisade.rule.Rule;
 import uk.gov.gchq.palisade.rule.Rules;
-import uk.gov.gchq.palisade.service.policy.request.Policy;
+import uk.gov.gchq.palisade.service.request.Policy;
 
 import java.util.Arrays;
 import java.util.Collections;
